+++ src/dotnet/Datack.Web.Service/Services/Agents.cs
@@ -14,12 +14,14 @@ namespace Datack.Web.Service.Services
     {
         private readonly AgentRepository _agentRepository;
         private readonly JobTaskRepository _jobTaskRepository;
+        private readonly JobRunTaskRepository _jobRunTaskRepository;
         private readonly RemoteService _remoteService;
 
-        public Agents(AgentRepository agentRepository, JobTaskRepository jobTaskRepository, RemoteService remoteService)
+        public Agents(AgentRepository agentRepository, JobTaskRepository jobTaskRepository, JobRunTaskRepository jobRunTaskRepository, RemoteService remoteService)
         {
             _agentRepository = agentRepository;
             _jobTaskRepository = jobTaskRepository;
+            _jobRunTaskRepository = jobRunTaskRepository;
             _remoteService = remoteService;
         }
 
@@ -151,9 +153,40 @@ namespace Datack.Web.Service.Services
         {
             var agent = await _agentRepository.GetById(agentId, cancellationToken);
 
+            if (agent == null)
+            {
+                throw new Exception($"Agent with ID {agentId} not found");
+            }
+
             var result = await _remoteService.GetLogs(agent, cancellationToken);
 
             return result;
+        }
+
+        public async Task UpgradeAgent(Guid agentId, CancellationToken cancellationToken)
+        {
+            var agent = await _agentRepository.GetById(agentId, cancellationToken);
+
+            if (agent == null)
+            {
+                throw new Exception($"Agent with ID {agentId} not found");
+            }
+
+            var jobRunTasks = await _jobRunTaskRepository.GetByAgentId(agentId, cancellationToken);
+
+            var jobRunsRunning = jobRunTasks.Any(m => m.JobRun.Completed == null);
+
+            if (jobRunsRunning)
+            {
+                throw new Exception($"Cannot upgrade agent, some jobs are running");
+            }
+
+            var result = await _remoteService.UpgradeAgent(agent, cancellationToken);
+
+            if (result != "Success")
+            {
+                throw new Exception("Failed to start upgrade");
+            }
         }
     }
 }