+++ src/react/src/pages/agents/AgentSummaryTab.tsx
@@ -1,12 +1,19 @@
-import { Table, Tbody, Td, Tr } from '@chakra-ui/react';
-import React, { FC } from 'react';
+import { Alert, AlertDescription, AlertIcon, Button, Table, Tbody, Td, Tr } from '@chakra-ui/react';
+import React, { FC, useState } from 'react';
+import useCancellationToken from '../../hooks/useCancellationToken';
 import { Agent } from '../../models/agent';
+import Agents from '../../services/agents';
 
 type Props = {
     agent: Agent;
 };
 
 const AgentSummaryTab: FC<Props> = (props) => {
+    const [error, setError] = useState<string | null>(null);
+    const [upgrading, setUpgrading] = useState<boolean>(false);
+
+    const cancelToken = useCancellationToken();
+
     const getState = (): string => {
         switch (props.agent.status) {
             case 'offline':
@@ -18,6 +25,17 @@ const AgentSummaryTab: FC<Props> = (props) => {
             default:
                 return props.agent.status || '';
         }
+    };
+
+    const handleUpgradeClick = async () => {
+        try {
+            setUpgrading(true);
+            await Agents.upgradeAgent(props.agent.agentId, cancelToken);
+        } catch (err: any) {
+            setError(err);
+        }
+
+        setUpgrading(false);
     };
 
     return (
@@ -40,9 +58,25 @@ const AgentSummaryTab: FC<Props> = (props) => {
                         <Td style={{ fontWeight: 'bold' }}>Version</Td>
                         <Td>{props.agent.version}</Td>
                     </Tr>
+                    <Tr>
+                        <Td></Td>
+                        <Td>
+                            {error != null ? (
+                                <Alert marginBottom={4} status="error">
+                                    <AlertIcon />
+                                    <AlertDescription>{error}</AlertDescription>
+                                </Alert>
+                            ) : null}
+                            <Button onClick={() => handleUpgradeClick()} isLoading={upgrading}>
+                                Upgrade Agent
+                            </Button>
+                        </Td>
+                    </Tr>
                 </Tbody>
             </Table>
-            <a href={`/Api/Agents/Logs/${props.agent.agentId}`} target="_blank" rel="noreferrer">View last 100 lines of agent log</a>
+            <a href={`/Api/Agents/Logs/${props.agent.agentId}`} target="_blank" rel="noreferrer">
+                View last 100 lines of agent log
+            </a>
         </>
     );
 };