+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 2021-11-01
+### Added
+- Add update button for agents
+
+## [1.0.3] - 2021-11-01
+### Added
+- Add HTTPS redirection
+
+### Changed
+- Add item select on the job run button
+- Some fixed to the update scripts
+
 ## [1.0.2] - 2021-10-31
 ### Changed
 - Maintenance build